API/requirements.txt
Pillow==9.4.0
Django==4.1.6
djangorestframework==3.14.0
django-extensions==3.2.1
pika==1.3.1